+++ packages/koop-core/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Only cache if geojson.ttl is set
+
 ## [4.0.3] - 2020-10-19
 ### Changed
 * Bump output-geoservices